backend-python/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
bcrypt==4.1.2
marshmallow==3.20.1
python-dotenv==1.0.0
